# portals/common/core/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "apk-site Admin / Release / Ops API",
    "version": "2.0.0",
    "description": "Auto-generated from Flask url_map (admin portal mode)."
  },
  "paths": {
...
    "/admin/projects/{project_id}/actions": {
      "get": {
        "summary": "project ops.ops platform actions page",
        "operationId": "project_ops.ops_platform_actions_page_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/admin/projects/{project_id}/agent-detail": {
      "get": {
        "summary": "project ops.ops platform agent detail page",
        "operationId": "project_ops.ops_platform_agent_detail_page_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/admin/projects/{project_id}/agent-device-local": {
      "get": {
        "summary": "project ops.ops platform agent device local page",
        "operationId": "project_ops.ops_platform_agent_device_local_page_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/admin/projects/{project_id}/agents": {
      "get": {
        "summary": "project ops.ops platform agent control page",
        "operationId": "project_ops.ops_platform_agent_control_page_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/admin/projects/{project_id}/change-governance": {
      "get": {
        "summary": "project ops.ops platform change governance page",
        "operationId": "project_ops.ops_platform_change_governance_page_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/admin/projects/{project_id}/diagnostics": {
      "get": {
        "summary": "project ops.ops platform diagnostics page",
        "operationId": "project_ops.ops_platform_diagnostics_page_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/admin/projects/{project_id}/ops": {
      "get": {
        "summary": "project ops.ops platform page",
        "operationId": "project_ops.ops_platform_page_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/admin/projects/{project_id}/topologies": {
      "get": {
        "summary": "project ops.ops platform topology page",
        "operationId": "project_ops.ops_platform_topology_page_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/admin/projects/{project_id}/topologies/canvas": {
      "get": {
        "summary": "project ops.ops platform topology canvas page",
        "operationId": "project_ops.ops_platform_topology_canvas_page_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/admin/projects/{project_id}/topology-bindings": {
      "get": {
        "summary": "project ops.project topology bindings page",
        "operationId": "project_ops.project_topology_bindings_page_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  }
}